Revisions and New GPC
The SoE (Lou Tyson) used the DOCX file to make formatting and typographical changes. Byron Villacorta is the new GPC, so the GPC contact information was updated. The PDF was generated from the DOCX file, and the PDF is now available on the chemical.olemiss.edu website.

The source MD file is still at version 0.1.0 and still needs to be updated to match the changes made here by Lou.